Use a composable to poll content tasks on all relevant pages.
